pyproject.toml
[tool.poetry]
name = "metadog"
version = "0.1.0"
description = ""
authors = ["Henning <henning.holgersen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fsspec = "^2023.6"
sqlalchemy = "^1.0"
paramiko = "^3.2"
smart-open = "^6.3"
pyyaml = "^6.0"
python-dotenv = "^1.0"
jinja2 = "^3.1"
click = "^8.1.3"
pandas = "^1.5"
scipy = "^1.10"

[project.optional-dependencies]
snowflake = [
    "snowflake-sqlalchemy ~=1.4.0"
]
azure = [
    "azure-identity ~= 1.12.0",
    "azure-storage-blob ~= 12.14.1"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
metadog = 'metadog.cli:metadog'